user login&user registration

+++ app/Http/Controllers/TasksController.php
@@ -4,22 +4,34 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-use App\Task;
+use App\Task; 
 
 class TasksController extends Controller
 {
+    
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
+    // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = Task::all()->where('user_id',"=" ,$user->id);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+        }
+        
+        return view('welcome', $data);
     }
 
     /**
@@ -27,10 +39,11 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
-        $task = new Task;
-        
+        $task = new task;
+
         return view('tasks.create', [
             'task' => $task,
         ]);
@@ -42,36 +55,43 @@ class TasksController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+    // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
-        //バリデーション
-        $request->validate([
-            'status' => 'required|max:10',
-            'content' => 'required|max:255',
+        $this->validate($request, [
+            'status' => 'required|max:10',   // 追加
+            'content' => 'required|max:191',
         ]);
-        
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
+
+        $task= new Task;
+        $task->status= $request -> status;
+        $task->content= $request ->content;
+        $task->user_id= \Auth::user()->id;
         $task->save();
         
         return redirect('/');
     }
-
+    
     /**
      * Display the specified resource.
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // getでtasks/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        $task = Task::findOrFail($id);
+        $task = Task::find($id);
         
-        return view('tasks.show', [
-           'task' => $task, 
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
+
 
     /**
      * Show the form for editing the specified resource.
@@ -79,13 +99,18 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // getでmessages/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
-        
-        return view('tasks.edit', [
-           'task' => $task, 
-        ]);
+        $task = Task::find($id);
+
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        } else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -95,19 +120,21 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'status' => 'required|max:10',   
-            'content' => 'required|max:255',
+        $this->validate($request, [
+            'status' => 'required|max:10',   // 追加
+            'content' => 'required|max:191',
         ]);
-        
-        $task = Task::findOrFail($id);
-        
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
+
+        $task = Task::find($id);
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        }
+
         return redirect('/');
     }
 
@@ -117,12 +144,14 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    // deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        
-        $task->delete();
-        
+        $task = \App\Task::find($id);
+
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         return redirect('/');
     }
 }
